Add auth.py views file
